abstracted logic to container
